+++ eventmesh-storage/eventmesh-storage-rabbitmq/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.consumer.Consumer
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-rabbitmq=org.apache.eventmesh.connector.rabbitmq.consumer.RabbitmqConsumer
+rabbitmq=org.apache.eventmesh.storage.rabbitmq.consumer.RabbitmqConsumer
